+++ dto/revocation_request.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019 Authlete, Inc.
+// Copyright (C) 2019-2021 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -25,4 +25,10 @@ type RevocationRequest struct {
 
 	//
 	ClientSecret string `json:"clientSecret,omitempty"`
+
+	//
+	ClientCertificate string `json:"clientCertificate,omitempty"`
+
+	//
+	ClientCertificatePath []string `json:"clientCertificatePath,omitempty"`
 }